+++ pkg/provider/provider.go
@@ -168,6 +168,7 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 		NewPlaygroundResource,
 		NewLLMBlueprintResource,
 		NewCustomModelResource,
+		NewCustomModelFromVectorDatabaseResource,
 		NewCustomModelLLMValidationResource,
 		NewCustomJobResource,
 		NewCustomMetricJobResource,
@@ -200,6 +201,7 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 		NewMemorySpaceResource,
 		NewArtifactResource,
 		NewWorkloadResource,
+		NewQuotaResource,
 	}
 }